Verify token hash from query string before passing onto confirmation
- Allow all default confirmations(text, page and redirect) to work
- Ensure query strings are using latest values from database

Token is a randomly generated uuid(v4).
Token hash is generated with `hash_hmac` using `sha512` as algo and
`wp_salt('auth')` as key.

Token hashes are valid **forever**.

Although we now confident that Gravity Forms is using the latest fields
values to build query strings, the token hash isn't an authentication
solution. It doesn't ensure the current visitor is the entry author.

Take just giving form as an example:
With ["using confirmation query strings to populate a form based on
another
submission"](https://docs.gravityforms.com/using-confirmation-query-strings-to-populate-a-form-based-on-another-submission/):
1. User fills in formA
1. User complete SagePay checkout form
1. User comes back and hits `CallbackHandler`
1. `CallbackHandler` sends user to formB according to confirmation
settings
1. User arrives formB url with merged query strings

Even with this pull request, Gravity Forms has no guarantee that user in
step 1 is the same person in the last step.

Possible Solution:
- Require WP user account registration or login in formA
- Gravity Forms auto-save user id during form submission
- Check [`wp_get_current_user`](https://codex.wordpress.org/Function_Reference/wp_get_current_user)
before echoing formB

Related: #12