Update `Plugin URI` to GitHub repo URL
Fix wp.org plugini validation error:
```
Error: Your plugin and author URIs are the same. A plugin URI (Uniform Resource Identifier) is a webpage that provides details about this specific plugin. An author URI is a webpage that provides information about the author of the plugin. Those two URIs must be different. You are not required to provide both, so pick the one that best applies to your situation.
```